% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add1tester.R
\name{add1tester}
\alias{add1tester}
\title{Increment by one using compiled C}
\usage{
add1tester(x)
}
\arguments{
\item{x}{Integer vector}
}
\value{
Integer vector incremented by one
}
\description{
Increment by one using compiled C
}
